# docs/openapi.yaml
openapi: 3.1.0
info:
  title: better-books API
  version: v1
  license:
    name: Proprietary — uso interno, não distribuído sob licença open source
  description: >
    Rede social para leitores — registro de leitura, avaliação, follow com aprovação
    e feed de atividade. Este documento consolida, num único contrato organizado por
    domínio de negócio, os 8 OpenAPI que antes viviam um por feature em
    `specs/00N-*/contracts/*.openapi.yaml` (001 a 008). Ele é a fonte única da verdade
    para quem consome a API — não é necessário voltar aos contratos antigos.

    Todo endpoint (exceto os marcados como públicos abaixo) exige
    `Authorization: Bearer <accessToken>`. Ver `auth-guide.md` para como obter e renovar
    o token. Toda lista paginada usa cursor opaco — ver `pagination-guide.md`. Toda
    resposta de erro segue o mesmo envelope — ver `error-catalog.md`. Um erro à parte,
    `DATABASE_UNAVAILABLE` (503), pode ocorrer em **qualquer** endpoint que dependa do
    MongoDB e por isso não é repetido em cada operação abaixo — só no catálogo.

    ## Notas de consolidação (RF-002, RF-020 — divergências corrigidas a favor do código)

    - `FeedItem` ganhou `reactionsCount` e `hasReacted` — presentes na implementação
      (`src/services/feed/types.ts`) mas ausentes no `feed.openapi.yaml` original (006),
      escrito antes de a feature 007-interactions adicionar reações.
    - `ReadingSession` inclui `review` em **toda** resposta desse formato (iniciar leitura,
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /users/{userId}/activity:
    get:
      tags: [feed]
      operationId: listUserActivity
      summary: >
        Atividade recente de uma pessoa, mais recente primeiro, paginada por cursor.
        Só acessível se você segue essa pessoa (follow aprovado) ou é o próprio
        usuário; qualquer outro caso responde o mesmo `404 USER_NOT_FOUND` de
        `GET /users/{userId}`. Cada item tem o mesmo formato de `GET /feed`.
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: Página do cursor (vazia se a pessoa é acessível mas não tem atividade).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedCursorPage" }
              examples:
                comAtividade:
                  value:
                    items:
                      - id: "64b7f0c2e1a2b3c4d5e6a101"
                        type: "review_published"
                        createdAt: "2026-09-01T12:00:00.000Z"
                        actor: { userId: "64b7f0c2e1a2b3c4d5e6f002", handle: "bruno", displayName: "Bruno" }
                        book: { id: "64b7f0c2e1a2b3c4d5e6b001", title: "Duna", authors: ["Frank Herbert"], coverUrl: null }
                        readingSessionId: "64b7f0c2e1a2b3c4d5e6c001"
                        currentPage: null
                        review: { id: "64b7f0c2e1a2b3c4d5e6d001", sessionId: "64b7f0c2e1a2b3c4d5e6c001", rating: 5, text: "Melhor releitura do ano.", containsSpoiler: false, createdAt: "2026-09-01T12:00:00.000Z", updatedAt: "2026-09-01T12:00:00.000Z" }
                        reactionsCount: 2
                        hasReacted: false
                    nextCursor: null
                vazia:
                  value: { items: [], nextCursor: null }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/UserNotFound" }

  # ---------------------------------------------------------------- follows
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId: { name: userId, in: path, required: true, schema: { type: string } }
...
    Cursor: { name: cursor, in: query, required: false, schema: { type: string, minLength: 1 } }
    Limit: { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  responses:
    ValidationError:
      description: "`VALIDATION_ERROR` — corpo/querystring reprovado pelo schema `zod`."
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    Unauthenticated:
      description: "`UNAUTHENTICATED` (sem Bearer válido) ou `INVALID_ACCESS_TOKEN` (expirado/malformado/conta inexistente)."
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
    UserNotFound:
      description: >
        `USER_NOT_FOUND` — `userId` não corresponde a nenhum usuário, OU o usuário
        existe mas não é visível para você (ex.: `GET /users/{userId}/activity` sem
        follow aprovado). As duas situações respondem exatamente igual — não vaza a
        existência de perfil privado (P6). Nunca `403`.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, statusCode]
          properties:
            code: { type: string, pattern: "^[A-Z][A-Z0-9_]*$" }
            message: { type: string }
            statusCode: { type: integer, minimum: 400, maximum: 599 }
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [path, message]
                properties:
                  path: { type: string }
                  message: { type: string }

    # --- auth / profile ---
...
    Review:
      type: object
      additionalProperties: false
      required: [id, sessionId, rating, text, containsSpoiler, createdAt, updatedAt]
      properties:
        id: { type: string }
        sessionId: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: [string, "null"], maxLength: 2000, description: "O texto completo sempre vem — `containsSpoiler` é só uma flag; ocultar é responsabilidade do cliente." }
        containsSpoiler: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    # --- feed ---
    FeedActor:
      type: object
      additionalProperties: false
      required: [userId, handle, displayName]
      properties:
        userId: { type: string }
        handle: { type: string }
        displayName: { type: string }
    FeedBook:
      type: object
      additionalProperties: false
      required: [id, title, authors, coverUrl]
      properties:
        id: { type: string }
        title: { type: string }
        authors: { type: array, items: { type: string } }
        coverUrl: { type: [string, "null"] }
    FeedItem:
      type: object
      additionalProperties: false
      required: [id, type, createdAt, actor, book, readingSessionId, currentPage, review, reactionsCount, hasReacted]
      properties:
        id: { type: string }
        type: { type: string, enum: [started_reading, finished_reading, review_published, progress_update] }
        createdAt: { type: string, format: date-time }
        actor: { $ref: "#/components/schemas/FeedActor" }
        book: { $ref: "#/components/schemas/FeedBook" }
        readingSessionId: { type: string }
        currentPage: { type: [integer, "null"], description: "Preenchido só quando `type === 'progress_update'`." }
        review:
          description: "Preenchido (estado atual, ao vivo) só quando `type === 'review_published'`."
          oneOf:
            - { $ref: "#/components/schemas/Review" }
            - type: "null"
        reactionsCount: { type: integer, minimum: 0 }
        hasReacted: { type: boolean, description: Se o usuário autenticado (viewer) já curtiu este item. }
    FeedCursorPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/FeedItem" } }
        nextCursor: { type: [string, "null"] }

    # --- comments ---
...